BD unsigned long long fix

+++ Dudar/tasks/BD.cpp
@@ -5,7 +5,7 @@ int main(){
     int N; //(5<= N <= 500)
     cin>>N;
 
-    int d[N+1][N+1];
+    unsigned long long int d[N+1][N+1];
 
     for(int i=0;i<N+1; i++){
         for(int j=0;j<N+1; j++) d[i][j]=0;
@@ -15,7 +15,7 @@ int main(){
     d[2][2] = 1;
 
 
-int count = 0;
+unsigned long long int count = 0;
     for(int i=3; i<N+1;i++){
         for(int j=1; j<i; j++){   
             count = 0;
@@ -25,7 +25,7 @@ int count = 0;
         d[i][i] = 1;
     }
 
-    int answer = 0;
+    unsigned long long int answer = 0;
 
     for(int i=1; i<N; i++) answer+=d[N][i];
     //answer--;
